--- src/models/connection/plugins/statics/index.js
@@ -1,6 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  findOwnConnections: require('./find-own-connections'),
-  updateConnectionStatus: require('./update-connection-status')
-})
