--- src/models/conversation/plugins/statics/find-one-with-participants.js
@@ -1,18 +1,0 @@
-'use strict'
-
-function findOneWithParticipants ({ conversationId }) {
-  if (!conversationId) throw new Error('Missing required conversationId')
-
-  return this.find({ _id: conversationId })
-    .limit(1)
-    .select('subject startDate messages participants')
-    .populate({
-      path: 'participants',
-      select: 'username name avatarUrl'
-    })
-    .exec()
-}
-
-exports = module.exports = function (schema, options) {
-  schema.statics.findOneWithParticipants = findOneWithParticipants
-}
