--- src/models/conversation/plugins/statics/index.js
@@ -1,7 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  findAllWithParticipants: require('./find-all-with-participants'),
-  findOneWithParticipants: require('./find-one-with-participants'),
-  updateUnreadStatus: require('./update-unread-status')
-})
