--- src/models/conversation/plugins/statics/update-unread-status.js
@@ -1,12 +1,0 @@
-'use strict'
-
-function updateUnreadStatus ({ target, status }) {
-  const updates = { '$set': { 'unread': status } }
-  const options = { new: true }
-
-  return this.findOneAndUpdate(target, updates, options).exec()
-}
-
-exports = module.exports = function (schema, options) {
-  schema.statics.updateUnreadStatus = updateUnreadStatus
-}
