--- src/models/message/plugins/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  statics: require('./statics')
-})
