--- src/models/message/plugins/statics/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  findByConversationAndRead: require('./find-by-conversation-and-read')
-})
