+++ src/models/post/index.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const mongoose = require('mongoose')
-const { statics } = require('./plugins')
 
 const postSchema = new mongoose.Schema({
   active: { type: Boolean, default: true },
@@ -27,14 +26,50 @@ const postSchema = new mongoose.Schema({
   createdAt: { type: Date }
 })
 
-// plug in static class methods
+/* ================================ METHODS ================================ */
 
-postSchema.plugin(statics.findPosts)
-postSchema.plugin(statics.updatePost)
-postSchema.plugin(statics.incrementViews)
-postSchema.plugin(statics.deletePost)
-postSchema.plugin(statics.deletePostsByAuthor)
+postSchema.statics.findPosts = function findPosts ({ filter }) {
+  return this.find(filter)
+    .populate('author', 'username name avatarUrl time_zone languages gender')
+    .exec()
+}
 
-// exports
+postSchema.statics.updatePost = function updatePost ({ target, updates, options = {} }) {
+  if (target == null) throw new Error('Missing required target param')
+  if (updates == null) throw new Error('Missing required updates param')
+  return this.findOneAndUpdate(target, updates, options).exec()
+}
+
+postSchema.statics.incrementViews = function incrementViews ({ target }) {
+  if (target == null) throw new Error('Missing required target param')
+
+  const updates = { $inc: { 'meta.views': 1 } }
+  return this.findOneAndUpdate(target, updates).exec()
+}
+
+/**
+ * Deletes a post by updating it to inactive
+ * @param  {Object}  target  Consisting of post _id & author
+ */
+postSchema.statics.deletePost = function deletePost ({ target }) {
+  if (target == null) throw new Error('Missing required target param')
+  const updates = {
+    deleted: true,
+    active: false,
+    updatedAt: new Date().toISOString()
+  }
+  const options = { new: true }
+  return this.findOneAndUpdate(target, updates, options).exec()
+}
+
+postSchema.statics.deletePostsByAuthor = function deletePostsByAuthor (authorId) {
+  const updates = { deleted: true, active: false }
+  const options = { multi: true }
+
+  return this.update({ author_id: authorId }, updates, options).exec()
+    .then(({ nModified }) => nModified > 0)
+}
+
+/* ================================ EXPORT ================================= */
 
 module.exports = mongoose.model('Post', postSchema)
