--- src/models/post/plugins/statics/find-posts.js
@@ -1,11 +1,0 @@
-'use strict'
-
-function findPosts ({ filter }) {
-  return this.find(filter)
-    .populate('author', 'username name avatarUrl time_zone languages gender')
-    .exec()
-}
-
-exports = module.exports = function (schema, options) {
-  schema.statics.findPosts = findPosts
-}
