--- src/models/post/plugins/statics/index.js
@@ -1,9 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  findPosts: require('./find-posts'),
-  updatePost: require('./update-post'),
-  incrementViews: require('./increment-post-views'),
-  deletePost: require('./delete-post'),
-  deletePostsByAuthor: require('./delete-posts-by-author')
-})
