--- src/models/post/plugins/statics/update-post.js
@@ -1,11 +1,0 @@
-'use strict'
-
-function updatePost ({ target, updates, options = {} }) {
-  if (target == null) throw new Error('Missing required target param')
-  if (updates == null) throw new Error('Missing required updates param')
-  return this.findOneAndUpdate(target, updates, options).exec()
-}
-
-exports = module.exports = function (schema, options) {
-  schema.statics.updatePost = updatePost
-}
