--- src/models/user/plugins/index.js
@@ -1,6 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  methods: require('./methods'),
-  statics: require('./statics')
-})
