--- src/models/user/plugins/methods/index.js
@@ -1,7 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  hashPassword: require('./hash-password'),
-  validatePassword: require('./validate-password'),
-  generateJwt: require('./generate-jwt')
-})
