--- src/models/user/plugins/statics/find-all-users.js
@@ -1,9 +1,0 @@
-'use strict'
-
-function findAllUsers ({ projection = {} }) {
-  return this.find({}, projection).exec()
-}
-
-exports = module.exports = function (schema, options) {
-  schema.statics.findAllUsers = findAllUsers
-}
