--- src/models/user/plugins/statics/index.js
@@ -1,7 +1,0 @@
-'use strict'
-
-exports = module.exports = Object.freeze({
-  findAllUsers: require('./find-all-users'),
-  updateUser: require('./update-user'),
-  deleteUser: require('./delete-user')
-})
